requirements.txt
streamlit
transformers
json
os
requests
torch
PyPDF2
pandas
python-dotenv
sentencepiece